ch: migrate UI to communicating to firebase through service layer
