+++ flutternotes/lib/services/auth/auth_provider.dart
@@ -1,6 +1,7 @@
 import 'package:flutternotes/services/auth/auth_user.dart';
 
 abstract class AuthProvider {
+  Future<void> initialize();
   AuthUser? get currentUser;
   Future<AuthUser> logIn({required String email, required String password});
   Future<AuthUser> createUser(
