+++ flutternotes/lib/services/auth/firebase_auth_provider.dart
@@ -1,4 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutternotes/firebase_options.dart';
 import 'package:flutternotes/services/auth/auth_exceptions.dart';
 import 'package:flutternotes/services/auth/auth_provider.dart';
 import 'package:flutternotes/services/auth/auth_user.dart';
@@ -87,5 +89,11 @@ class FirebaseAuthProvider implements AuthProvider {
     } else {
       throw UserNotLoggedInuthExceptions();
     }
+  }
+
+  @override
+  Future<void> initialize() async {
+    await Firebase.initializeApp(
+        options: DefaultFirebaseOptions.currentPlatform);
   }
 }
