+++ flutternotes/lib/views/notes_view.dart
@@ -1,0 +1,66 @@
+import 'package:flutter/material.dart';
+import 'package:flutternotes/services/auth/auth_service.dart';
+
+import '../constants/routes.dart';
+import '../enums/menu_action.dart';
+
+class NotesView extends StatefulWidget {
+  const NotesView({Key? key}) : super(key: key);
+
+  @override
+  State<NotesView> createState() => _NotesViewState();
+}
+
+class _NotesViewState extends State<NotesView> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Main UI'),
+        actions: [
+          PopupMenuButton<MenuAction>(
+            onSelected: (value) async {
+              switch (value) {
+                case MenuAction.logout:
+                  final shouldLogout = await showLogOutDialog(context);
+                  if (shouldLogout) {
+                    await AuthService.firebase().logOut();
+                    Navigator.of(context)
+                        .pushNamedAndRemoveUntil(loginRoute, (_) => false);
+                  }
+                  break;
+              }
+
+              ;
+            },
+            itemBuilder: (context) {
+              return const [
+                PopupMenuItem(value: MenuAction.logout, child: Text('logout'))
+              ];
+            },
+          )
+        ],
+      ),
+      body: const Text('hello World'),
+    );
+  }
+}
+
+Future<bool> showLogOutDialog(BuildContext context) {
+  return showDialog<bool>(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text('Sign Out'),
+          content: const Text('Are you sure you want to sign out?'),
+          actions: [
+            TextButton(
+                onPressed: () => {Navigator.of(context).pop(false)},
+                child: const Text('Cancel')),
+            TextButton(
+                onPressed: () => {Navigator.of(context).pop(true)},
+                child: const Text('Log Out'))
+          ],
+        );
+      }).then((value) => value ?? false);
+}
