+++ flutternotes/lib/views/register_view.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutternotes/constants/routes.dart';
+import 'package:flutternotes/services/auth/auth_exceptions.dart';
+import 'package:flutternotes/services/auth/auth_service.dart';
 
 import '../utilities/show_error_dialog.dart';
 
@@ -50,36 +52,22 @@ class _RegisterViewState extends State<RegisterView> {
         ElevatedButton(
           onPressed: () async {
             try {
-              final usercredential = await FirebaseAuth.instance
-                  .createUserWithEmailAndPassword(
-                      email: email.text, password: password.text);
-              final user = FirebaseAuth.instance.currentUser;
-              await user?.sendEmailVerification();
+              await AuthService.firebase()
+                  .createUser(email: email.text, password: password.text);
+
+              AuthService.firebase().currentUser;
+
+              await AuthService.firebase().sendEmailVerification();
+
               Navigator.of(context).pushNamed(verifyEmailRoute);
-            } on FirebaseAuthException catch (e) {
-              if (e.code == 'weak-password') {
-                await showErrorDialog(
-                  context,
-                  'Weak Password',
-                );
-              } else if (e.code == 'email-already-in-use') {
-                await showErrorDialog(
-                  context,
-                  'Email is already in use',
-                );
-              } else if (e.code == 'invalid-email') {
-                await showErrorDialog(
-                  context,
-                  'This is an invalid email',
-                );
-              } else {
-                await showErrorDialog(
-                  context,
-                  'Error ${e.code}',
-                );
-              }
-            } catch (e) {
-              await showErrorDialog(context, e.toString());
+            } on WeakPasswordAuthException {
+              await showErrorDialog(context, 'Weak Password');
+            } on EmailAlreadyInUseAuthException {
+              await showErrorDialog(context, 'Email already in use');
+            } on InvalidEmailAuthException {
+              await showErrorDialog(context, 'Invalid email');
+            } on GenericAuthException {
+              await showErrorDialog(context, 'Failed to register');
             }
           },
           child: const Text('Register'),
